cleanup: Expose `struct Tox` to internal code.
To avoid unsafe casts that tokstyle doesn't allow and that can cause
random breakages.
