+++ auto_tests/TCP_test.c
@@ -761,7 +761,7 @@ static bool tcp_oobdata_callback_called;
 static int tcp_oobdata_callback(void *object, const uint8_t *public_key, unsigned int id, const uint8_t *data,
                                 uint16_t length, void *userdata)
 {
-    TCP_Connections *tcp_c = (TCP_Connections *)object;
+    const TCP_Connections *tcp_c = (const TCP_Connections *)object;
 
     if (length != 6) {
         return -1;
