+++ auto_tests/lan_discovery_test.c
@@ -3,6 +3,7 @@
 
 #include "../testing/misc_tools.h"
 #include "../toxcore/ccompat.h"
+#include "../toxcore/tox_struct.h"
 #include "auto_test_support.h"
 
 static uint64_t get_state_clock_callback(Mono_Time *mono_time, void *user_data)
@@ -20,16 +21,13 @@ int main(void)
     ck_assert(tox1 != nullptr);
     ck_assert(tox2 != nullptr);
 
-    // TODO(iphydf): Don't rely on toxcore internals.
-    uint64_t clock = current_time_monotonic(((Messenger *)tox1)->mono_time);
+    uint64_t clock = current_time_monotonic(tox1->mono_time);
     Mono_Time *mono_time;
 
-    // TODO(iphydf): Don't rely on toxcore internals.
-    mono_time = ((Messenger *)tox1)->mono_time;
+    mono_time = tox1->mono_time;
     mono_time_set_current_time_callback(mono_time, get_state_clock_callback, &clock);
 
-    // TODO(iphydf): Don't rely on toxcore internals.
-    mono_time = ((Messenger *)tox2)->mono_time;
+    mono_time = tox2->mono_time;
     mono_time_set_current_time_callback(mono_time, get_state_clock_callback, &clock);
 
     printf("Waiting for LAN discovery. This loop will attempt to run until successful.");
