+++ other/bootstrap_daemon/docker/tox-bootstrapd.sha256
@@ -1,1 +1,1 @@
-c93318d8f05a4e9c9bd74bfb6f01b776c7dcb7450745dfe5f31dac592353cd74  /usr/local/bin/tox-bootstrapd
+41b304348d45a0389807fa9bf45fec05b1c34e47905656923fb3cbc172c2c9ff  /usr/local/bin/tox-bootstrapd
