+++ toxav/toxav_old.c
@@ -8,16 +8,12 @@
  */
 #include "toxav.h"
 
+#include "../toxcore/tox_struct.h"
 #include "groupav.h"
 
 int toxav_add_av_groupchat(Tox *tox, audio_data_cb *audio_callback, void *userdata)
 {
-    // TODO(iphydf): Don't rely on toxcore internals.
-    const Messenger *m;
-    //!TOKSTYLE-
-    m = *(const Messenger **)tox;
-    //!TOKSTYLE+
-    return add_av_groupchat(m->log, tox, m->conferences_object, audio_callback, userdata);
+    return add_av_groupchat(tox->m->log, tox, tox->m->conferences_object, audio_callback, userdata);
 }
 
 /** @brief Join a AV group (you need to have been invited first).
@@ -30,12 +26,7 @@ int toxav_add_av_groupchat(Tox *tox, audio_data_cb *audio_callback, void *userda
 int toxav_join_av_groupchat(Tox *tox, uint32_t friendnumber, const uint8_t *data, uint16_t length,
                             audio_data_cb *audio_callback, void *userdata)
 {
-    // TODO(iphydf): Don't rely on toxcore internals.
-    const Messenger *m;
-    //!TOKSTYLE-
-    m = *(const Messenger **)tox;
-    //!TOKSTYLE+
-    return join_av_groupchat(m->log, tox, m->conferences_object, friendnumber, data, length, audio_callback, userdata);
+    return join_av_groupchat(tox->m->log, tox, tox->m->conferences_object, friendnumber, data, length, audio_callback, userdata);
 }
 
 /** @brief Send audio to the group chat.
@@ -55,12 +46,7 @@ int toxav_join_av_groupchat(Tox *tox, uint32_t friendnumber, const uint8_t *data
 int toxav_group_send_audio(Tox *tox, uint32_t groupnumber, const int16_t *pcm, unsigned int samples, uint8_t channels,
                            uint32_t sample_rate)
 {
-    // TODO(iphydf): Don't rely on toxcore internals.
-    const Messenger *m;
-    //!TOKSTYLE-
-    m = *(const Messenger **)tox;
-    //!TOKSTYLE+
-    return group_send_audio(m->conferences_object, groupnumber, pcm, samples, channels, sample_rate);
+    return group_send_audio(tox->m->conferences_object, groupnumber, pcm, samples, channels, sample_rate);
 }
 
 /** @brief Enable A/V in a groupchat.
@@ -80,12 +66,7 @@ int toxav_group_send_audio(Tox *tox, uint32_t groupnumber, const int16_t *pcm, u
  */
 int toxav_groupchat_enable_av(Tox *tox, uint32_t groupnumber, audio_data_cb *audio_callback, void *userdata)
 {
-    // TODO(iphydf): Don't rely on toxcore internals.
-    const Messenger *m;
-    //!TOKSTYLE-
-    m = *(const Messenger **)tox;
-    //!TOKSTYLE+
-    return groupchat_enable_av(m->log, tox, m->conferences_object, groupnumber, audio_callback, userdata);
+    return groupchat_enable_av(tox->m->log, tox, tox->m->conferences_object, groupnumber, audio_callback, userdata);
 }
 
 /** @brief Disable A/V in a groupchat.
@@ -95,21 +76,11 @@ int toxav_groupchat_enable_av(Tox *tox, uint32_t groupnumber, audio_data_cb *aud
  */
 int toxav_groupchat_disable_av(Tox *tox, uint32_t groupnumber)
 {
-    // TODO(iphydf): Don't rely on toxcore internals.
-    const Messenger *m;
-    //!TOKSTYLE-
-    m = *(const Messenger **)tox;
-    //!TOKSTYLE+
-    return groupchat_disable_av(m->conferences_object, groupnumber);
+    return groupchat_disable_av(tox->m->conferences_object, groupnumber);
 }
 
 /** @brief Return whether A/V is enabled in the groupchat. */
 bool toxav_groupchat_av_enabled(Tox *tox, uint32_t groupnumber)
 {
-    // TODO(iphydf): Don't rely on toxcore internals.
-    const Messenger *m;
-    //!TOKSTYLE-
-    m = *(const Messenger **)tox;
-    //!TOKSTYLE+
-    return groupchat_av_enabled(m->conferences_object, groupnumber);
+    return groupchat_av_enabled(tox->m->conferences_object, groupnumber);
 }
