+++ toxcore/network.c
@@ -679,7 +679,7 @@ non_null()
 static int receivepacket(const Logger *log, Socket sock, IP_Port *ip_port, uint8_t *data, uint32_t *length)
 {
     memset(ip_port, 0, sizeof(IP_Port));
-    struct sockaddr_storage addr;
+    struct sockaddr_storage addr = {0};
 #ifdef OS_WIN32
     int addrlen = sizeof(addr);
 #else
